feat (frontend): add Workflow Validator util for invalid config indication, refactor WorkflowBuilder
- Added node types management in a separate node-types file.
- Implemented workflow validation logic in a new utility.
